reaperbot/pyproject.toml
[project]
name = "reaperbot"
version = "0.1.0"
description = "ReaperBot Agentic AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "devtools>=0.12.2",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "jinja2>=3.1.5",
    "openai>=1.61.1",
    "pydantic-ai>=0.0.22",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "rich>=13.9.4",
    "tenacity>=9.0.0",
    "uvicorn[standard]>=0.33.0",
]
